refactor: alerts context provider
